implementing user activation and some refactory

+++ src/main/java/br/com/personalog/dao/MoodDao.java
@@ -4,5 +4,5 @@ import br.com.personalog.dao.query.MoodDaoQuery;
 import br.com.personalog.model.Mood;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MoodDao extends JpaRepository<Mood, Integer>, MoodDaoQuery  {
+public interface MoodDao extends JpaRepository<Mood, Long>, MoodDaoQuery  {
 }
